services/service1/requirements.txt
uvicorn
fastapi
pydantic
python-dotenv
pytest
httpx
uuid